Update README.md
Increasing aridity and excessive water withdrawal from groundwater bodies leads to a decline of groundwater levels. Where and why groundwater levels vary has recently been studied by CORRECTIV: Wo in Deutschland das Grundwasser sinkt (in german language). CORRECTIV is an independent Organisation for investigative journalism and research.
The data of this project is a slightly revised analysis of the data provided by CORRECTIV on their GitHub repository. Based on groundwater levels measured in Brandenburg in June, July and August, a trend [m/year] was derived for the last 5 years. Negative values refer to declining groundwater levels and positive values to rising levels. Note that there may be duplicate observations (groundwater stations with the same coordinates).
The aim of the project is to identify spatial controls (spatial trends or covariates) on and possible interactions of observed groundwater trends